add load function to lazy export minimal package code
